refactored.  Redux integration not started
