+++ package.json
@@ -4,9 +4,10 @@
   "private": true,
   "homepage": "./",
   "devDependencies": {
+    "concurrently": "^3.4.0",
     "gh-pages": "^0.12.0",
     "react-scripts": "0.9.5",
-    "concurrently": "^3.4.0"
+    "redux-devtools": "^3.3.2"
   },
   "dependencies": {
     "axios": "^0.15.3",
@@ -16,6 +17,7 @@
     "react": "^15.4.2",
     "react-dom": "^15.4.2",
     "react-grid-system": "^2.7.0",
+    "react-redux": "^5.0.3",
     "react-tap-event-plugin": "^2.0.1",
     "redux": "^3.6.0"
   },
