+++ src/ReduxActions.js
@@ -1,0 +1,8 @@
+export const SET_STEP='SET_STEP';
+
+export const setStep=(step)=>{
+    return {
+        type:SET_STEP,
+        step
+    };
+}
